generalize pool and move to shared
