+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SNI/SNIPhysicalHandle.cs
@@ -15,11 +15,11 @@ namespace Microsoft.Data.SqlClient.SNI
 #if DEBUG
         private static int s_packetId;
 #endif
-        private SNIPacketPool _pool;
+        private Pool<SNIPacket> _pool;
 
         protected SNIPhysicalHandle(int poolSize = DefaultPoolSize)
         {
-            _pool = new SNIPacketPool(poolSize);
+            _pool = new Pool<SNIPacket>(poolSize);
         }
 
         public override SNIPacket RentPacket(int headerSize, int dataSize)
